export shared state from a single file
